+++ autochannel/__init__.py
@@ -1,5 +1,5 @@
 """
 init - version info
 """
-VERSION_INFO = (2, 4, 0)
+VERSION_INFO = (3, 0, 0)
 VERSION = '.'.join(str(c) for c in VERSION_INFO)
