+++ autochannel/autochannel.py
@@ -6,7 +6,7 @@ import aiohttp
 import json
 from discord.ext.commands import Bot
 """monitor"""
-from autochannel.lib.discordDog import DDAgent
+
 """data"""
 from autochannel.data.database import DB
 
@@ -26,12 +26,6 @@ class AutoChannel(discord.ext.commands.Bot):
         self.app_id = kwargs.get('app_id')
         self.auto_channel_prefix = kwargs.get('auto_channel_prefix')
         self.auto_categories = kwargs.get('auto_categories')
-        self.dd_api_key = kwargs.get('dd_api_key')
-        self.dd_app_key = kwargs.get('dd_app_key')
         self.env = kwargs.get('env') or 'dev'
         self.stats = None
         self.voice_channel_prefix = kwargs.get('voice_channel_prefix')
-
-        if self.dd_api_key and self.dd_app_key:
-           self.stats = DDAgent(dd_api_key=kwargs.get('dd_api_key'), dd_app_key=kwargs.get('dd_app_key'), constant_tags=[f'env:{self.env}', 'bot:autochannel'])
-           self.stats.increment('autochannel_bot.start.count')
