+++ autochannel/autochannel_bot.py
@@ -13,7 +13,8 @@ import sys
 from discord import Game
 from discord.ext import commands
 from discord.ext.commands import Bot
-
+"""metrics"""
+from prometheus_client import start_http_server
 """AC Imports"""
 from autochannel.lib import plugin, utils
 from autochannel.autochannel import AutoChannel
@@ -34,8 +35,6 @@ TOKEN = os.getenv('TOKEN')
 VOICE_CHANNEL_PREFIX = os.getenv('VOICE_CHANNEL_PREFIX') or '!VC '
 AUTO_CHANNEL_PREFIX = os.getenv('AUTO_CHANNEL_PREFIX') or '!AC '
 AUTO_CATEGORIES = os.getenv('AUTO_CATEGORIES').lower().split(",") or ['auto-voice']
-DD_API_KEY = os.getenv('DD_API_KEY') or None
-DD_APP_KEY = os.getenv('DD_APP_KEY') or None
 ENV = os.getenv('ENV') or None
 
 def main():
@@ -60,12 +59,13 @@ def main():
     autochannel = AutoChannel(shard_id=int(SHARD), shard_count=int(SHARD_COUNT),
                     command_prefix=BOT_PREFIX, app_id=APP_ID, voice_channel_prefix=VOICE_CHANNEL_PREFIX,
                     auto_channel_prefix=AUTO_CHANNEL_PREFIX, auto_categories=AUTO_CATEGORIES,
-                    dd_api_key=DD_API_KEY, dd_app_key=DD_APP_KEY, env=ENV)
+                    env=ENV)
 
     for extension in EXTENSIONS:
         plugin.load('autochannel.lib.plugins.{}'.format(extension), autochannel)
-
+    start_http_server(8000)
     autochannel.run(TOKEN)
+   
 
 
 if __name__ == '__main__':
