+++ setup.py
@@ -8,11 +8,11 @@ INSTALL_REQUIREMENTS = [
     'asyncio',
     'aiomeasures',
     'coloredlogs',
-    'datadog',
     'discord.py==1.2.3',
     'flask_sqlalchemy',
     'pip==18.0',
     'profanityfilter',
+    'prometheus_client',
     'psycopg2-binary',
     'pyyaml',
     'requests==2.21.0',
